python/pyproject.toml
[project]
name = "wifi-visualizer"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "paho-mqtt",
    "matplotlib",
]
